chore(PT-47): reorganize files into feature-based structure, fix typo in footer, update image placeholder
