+++ src/features/searchSpecialists/components/SpecialistCard.tsx
